Define the pydantic models for the user_profile table.
Some models needed to be defined as a class dependency because the string fields needed to be specified as a Form() field, as you cannot have string fields along with File() parameters.

This means that Skill needs to be created separately for a user as Skill is a pydantic model and will be interpreted as a request body and cannot be included in a request with form fields.

Signed-off-by: longfight123 <lingphoonlee@gmail.com>